requirements.txt
streamlit==1.46.0
langchain==0.3.26
langchain-community==0.3.26
langchain-groq==0.3.4
python-dotenv==1.1.1